Turtle Sensor 0.1.0/requirements.txt
pandas==0.19.2
pyserial==3.2.1
python-dateutil==2.6.0